backend/requirements.txt
fastapi==0.88.0
openai
pip
requests
uvicorn
python-dotenv